backend/requirements.txt
openai
moviepy
google-cloud
google-cloud-storage
google-api-python-client4
Flask
flask-cors